jurify-backend/requirements.txt
flask
flask-cors
torch
transformers
sentence-transformers
spacy
PyPDF2
python-docx
reportlab
nltkgunicorn
